move ProposePlaceService(remove jackson and swagger dependencies)
