+++ core/src/test/java/greencity/service/impl/PlaceServiceImplTest.java
@@ -9,7 +9,7 @@ import greencity.dto.filter.FilterPlaceDto;
 import greencity.dto.location.LocationAddressAndGeoDto;
 import greencity.dto.location.LocationVO;
 import greencity.dto.openhours.OpeningHoursDto;
-import greencity.dto.openinghours.OpeningHoursVO;
+import greencity.dto.openhours.OpeningHoursVO;
 import greencity.dto.photo.PhotoAddDto;
 import greencity.dto.place.*;
 import greencity.entity.Category;
