+++ service-api/src/main/java/greencity/dto/breaktime/BreakTimeDto.java
@@ -1,7 +1,5 @@
 package greencity.dto.breaktime;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-import io.swagger.annotations.ApiModelProperty;
 import java.time.LocalTime;
 import javax.validation.constraints.NotNull;
 import lombok.*;
@@ -13,12 +11,8 @@ import lombok.*;
 @EqualsAndHashCode(exclude = {"startTime", "endTime"})
 public class BreakTimeDto {
     @NotNull
-    @JsonFormat(pattern = "HH:mm", shape = JsonFormat.Shape.STRING)
-    @ApiModelProperty(dataType = "java.lang.String")
     private LocalTime startTime;
 
     @NotNull
-    @JsonFormat(pattern = "HH:mm", shape = JsonFormat.Shape.STRING)
-    @ApiModelProperty(dataType = "java.lang.String")
     private LocalTime endTime;
 }
