+++ service-api/src/main/java/greencity/dto/openhours/OpeningHoursDto.java
@@ -1,0 +1,25 @@
+package greencity.dto.openhours;
+
+import greencity.dto.breaktime.BreakTimeDto;
+import java.time.DayOfWeek;
+import java.time.LocalTime;
+import javax.validation.constraints.NotNull;
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@EqualsAndHashCode(exclude = {"openTime", "closeTime"})
+public class OpeningHoursDto {
+    @NotNull
+    private LocalTime openTime;
+
+    @NotNull
+    private LocalTime closeTime;
+
+    @NotNull
+    private DayOfWeek weekDay;
+
+    private BreakTimeDto breakTime;
+}
