--- service-api/src/main/java/greencity/dto/openinghours/OpenHoursDto.java
@@ -1,27 +1,0 @@
-package greencity.dto.openinghours;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import greencity.dto.breaktime.BreakTimeDto;
-import java.time.DayOfWeek;
-import java.time.LocalTime;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class OpenHoursDto implements Comparable<OpenHoursDto> {
-    private Long id;
-    @JsonFormat(pattern = "HH:mm")
-    private LocalTime openTime;
-    @JsonFormat(pattern = "HH:mm")
-    private LocalTime closeTime;
-    private DayOfWeek weekDay;
-    private BreakTimeDto breakTime;
-
-    @Override
-    public int compareTo(OpenHoursDto openHoursDto) {
-        return this.weekDay.compareTo(openHoursDto.weekDay);
-    }
-}
