--- service-api/src/main/java/greencity/dto/openinghours/OpeningHoursDto.java
@@ -1,32 +1,0 @@
-package greencity.dto.openinghours;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import greencity.dto.breaktime.BreakTimeDto;
-import io.swagger.annotations.ApiModelProperty;
-import lombok.*;
-
-import javax.validation.constraints.NotNull;
-import java.time.DayOfWeek;
-import java.time.LocalTime;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@EqualsAndHashCode(exclude = {"openTime", "closeTime"})
-public class OpeningHoursDto {
-    @NotNull
-    @JsonFormat(pattern = "HH:mm", shape = JsonFormat.Shape.STRING)
-    @ApiModelProperty(dataType = "java.lang.String")
-    private LocalTime openTime;
-
-    @NotNull
-    @JsonFormat(pattern = "HH:mm", shape = JsonFormat.Shape.STRING)
-    @ApiModelProperty(dataType = "java.lang.String")
-    private LocalTime closeTime;
-
-    @NotNull
-    private DayOfWeek weekDay;
-
-    private BreakTimeDto breakTime;
-}
