+++ service/src/test/java/greencity/ModelUtils.java
@@ -12,7 +12,7 @@ import greencity.dto.language.LanguageDTO;
 import greencity.dto.language.LanguageTranslationDTO;
 import greencity.dto.language.LanguageVO;
 import greencity.dto.location.LocationAddressAndGeoDto;
-import greencity.dto.openinghours.OpeningHoursDto;
+import greencity.dto.openhours.OpeningHoursDto;
 import greencity.dto.user.UserGoalResponseDto;
 import greencity.dto.user.UserGoalVO;
 import greencity.dto.user.UserVO;
